requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy
Click>=7.1.2
Flask>=1.1.2
itsdangerous==1.1.0
Jinja2>=2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
gunicorn>=20.0.4
torch==1.8.0+cpu
torchvision==0.9.0+cpu
python-dotenv==0.19.1
joblib==1.1.0
